result data generic type inferencing
