+++ qbraid/_version.py
@@ -14,4 +14,4 @@ Module containing version information
 Version number (major.minor.patch[-label])
 
 """
-__version__ = "0.8.7"
+__version__ = "0.8.8.dev"
